returning error messages and other minor refactoring
