+++ indra-client/src/main/java/org/lambda3/indra/client/NeighborsVectorsRequest.java
@@ -57,7 +57,11 @@ public class NeighborsVectorsRequest extends AbstractBasicRequest<NeighborsVecto
         checkAndAppendErrorMessagesLists(terms, "terms", errorMessages);
 
         if (topk <= 0) {
-            errorMessages.append(" - 'topk' must be greater than 0\n");
+            errorMessages.append(" - 'topk' must be greater than 0;\n");
+        }
+
+        if(isMt()) {
+            errorMessages.append(" - machine translation not supported in this function. 'mt' can't be true;\n");
         }
 
         return errorMessages.toString();
