+++ indra-service/src/main/java/org/lambda3/indra/service/Indra.java
@@ -26,7 +26,6 @@ package org.lambda3.indra.service;
  * ==========================License-End===============================
  */
 
-import org.lambda3.indra.service.impl.Server;
 import org.slf4j.bridge.SLF4JBridgeHandler;
 
 /**
