+++ indra-service/src/main/java/org/lambda3/indra/service/mapper/SerializationExceptionMapper.java
@@ -1,4 +1,4 @@
-package org.lambda3.indra.service.impl;
+package org.lambda3.indra.service.mapper;
 
 /*-
  * ==========================License-Start=============================
@@ -26,23 +26,23 @@ package org.lambda3.indra.service.impl;
  * ==========================License-End===============================
  */
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.inject.Singleton;
-import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
 import javax.ws.rs.ext.Provider;
 
 @Provider
 @Singleton
-public final class ErrorMapper implements ExceptionMapper<WebApplicationException> {
-    private static Logger logger = LoggerFactory.getLogger(ErrorMapper.class);
+public class SerializationExceptionMapper implements ExceptionMapper<JsonProcessingException> {
+    private static Logger logger = LoggerFactory.getLogger(SerializationExceptionMapper.class);
 
     @Override
-    public Response toResponse(WebApplicationException exception) {
-        logger.error("Web App Error!", exception);
-        return exception.getResponse();
+    public Response toResponse(JsonProcessingException exception) {
+        logger.error("Serialization Error", exception);
+        return Response.status(Response.Status.BAD_REQUEST).build();
     }
 }
