+++ indra-service/src/main/java/org/lambda3/indra/service/mock/MockedRelatednessResourceImpl.java
@@ -1,4 +1,4 @@
-package org.lambda3.indra.service.impl.mock;
+package org.lambda3.indra.service.mock;
 
 /*-
  * ==========================License-Start=============================
@@ -12,10 +12,10 @@ package org.lambda3.indra.service.impl.mock;
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * 
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- *
+ * 
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -26,32 +26,39 @@ package org.lambda3.indra.service.impl.mock;
  * ==========================License-End===============================
  */
 
-import org.apache.commons.math3.linear.ArrayRealVector;
-import org.apache.commons.math3.linear.RealVectorUtil;
-import org.apache.commons.math3.random.UnitSphereRandomVectorGenerator;
-import org.lambda3.indra.client.SparseVectorResponse;
-import org.lambda3.indra.client.VectorRequest;
-import org.lambda3.indra.client.VectorResource;
-import org.lambda3.indra.client.VectorResponse;
+import org.lambda3.indra.client.*;
 
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 
 /**
- * MockedVectorResourceImpl implementation that returns random vectors.
+ * RelatednessResource implementation that randomly assigns a relatedness value.
  * For testing puporses.
  */
-public class MockedVectorResourceImpl implements VectorResource {
+public class MockedRelatednessResourceImpl implements RelatednessResource {
 
-    private static final int NUM_DIMENSIONS = 300;
-    private static final UnitSphereRandomVectorGenerator rvg = new UnitSphereRandomVectorGenerator(NUM_DIMENSIONS);
+    private static Random rnd = new Random();
 
     @Override
-    public VectorResponse getVector(VectorRequest request) {
-        Map<String, Map<Integer, Double>> terms = new HashMap<>();
-        request.getTerms().forEach(t -> terms.put(t, RealVectorUtil.vectorToMap(new ArrayRealVector(rvg.nextVector()))));
-        VectorResponse response = new SparseVectorResponse(request, terms);
+    public RelatednessPairResponse getRelatedness(RelatednessPairRequest request) {
+        Collection<ScoredTextPair> scored = new ArrayList<>();
 
-        return response;
+        for (TextPair pair : request.getPairs()) {
+            AnalyzedPair analyzedPair = new AnalyzedPair(pair);
+            ScoredTextPair stp = new ScoredTextPair(analyzedPair, rnd.nextDouble());
+            scored.add(stp);
+        }
+
+        return new RelatednessPairResponse(request, scored);
+    }
+
+    @Override
+    public RelatednessOneToManyResponse getRelatedness(RelatednessOneToManyRequest request) {
+        Map<String, Double> results = new LinkedHashMap<>();
+
+        for (String m : request.getMany()) {
+            results.put(m, rnd.nextDouble());
+        }
+
+        return new RelatednessOneToManyResponse(request, results);
     }
 }
